requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pulp>=2.8.0
pandas>=2.2.0
numpy>=1.26.0
nba_api>=1.4.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2,<4.0
cryptography>=42.0.0
slowapi>=0.1.9
openpyxl>=3.1.0
